llamo-reimplement/requirements.txt
absl-py==2.1.0
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.0
aliyun-python-sdk-kms==2.16.2
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
arrow==1.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bigmodelvis==0.0.1
bleach==6.1.0
blessed==1.20.0
blinker==1.7.0
blis==0.7.11
boto3==1.35.16
botocore==1.35.16
braceexpand==0.1.7
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cheroot==10.0.0
click==8.1.7
cloudpathlib==0.16.0
cmake==3.28.3
comm==0.2.2
confection==0.1.4
contextlib2==21.6.0
contexttimer==0.3.3
contourpy==1.1.1
crcmod==1.7
croniter==1.4.1
cryptography==42.0.5
cycler==0.12.1
cymem==2.0.8
datasets==2.18.0
dateutils==0.6.12
debugpy==1.8.1
decorator==5.1.1
decord==0.6.0
deepdiff==7.0.1
deepspeed==0.14.0
delta-center-client==0.0.4
descriptastorus @ git+https://github.com/bp-kelley/descriptastorus@9b133e2c91bb6a67df53db4cba992776db219ab7
dgl==2.1.0
dill==0.3.8
distlib==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
editor==1.6.6
einops==0.7.0
exceptiongroup==1.2.1
executing==2.0.1
fairscale==0.4.4
fastapi==0.114.1
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.7
identify==2.5.35
idna==3.6
imageio==2.34.0
importlib_metadata==7.1.0
importlib_resources==6.1.3
inquirer==3.4.0
iopath==0.1.10
ipykernel==6.29.4
ipynb==0.5.1
ipython==8.12.3
itsdangerous==2.2.0
jaraco.functools==4.0.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kaggle==1.6.6
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
Levenshtein==0.25.1
lightning==2.0.7
lightning-cloud==0.5.70
lightning-utilities==0.10.1
lit==17.0.6
littleutils==0.2.2
lmdb==1.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-collections==0.1.1
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
nltk==3.8.1
nodeenv==1.8.0
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
ogb==1.3.5
omegaconf==2.3.0
openai==1.28.1
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
opendelta @ git+https://github.com/thunlp/OpenDelta.git@067eed2304cb1bdfe462094e42a37de4de98edff
ordered-set==4.1.0
oss2==2.15.0
outdated==0.2.2
packaging==23.2
pandas==2.0.3
pandas-flavor==0.6.0
parso==0.8.3
peft==0.9.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
plotly==5.19.0
portalocker==2.8.2
pre-commit==3.5.0
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
PubChemPy==1.0.4
pure-eval==0.2.2
py-cpuinfo==9.0.0
py2opsin==1.1.0
pyarrow==15.0.1
pyarrow-hotfix==0.6
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.1.1
pydantic_core==2.4.0
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.9.0
pynvml==11.5.0
pyparsing==3.1.2
pyproject-toml==0.0.10
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-multipart==0.0.9
python-slugify==8.0.4
pytorch-lightning==2.2.1
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.6
rdkit==2023.9.5
readchar==4.2.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rouge-score==0.1.2
rpds-py==0.18.0
runs==1.2.2
s3transfer==0.10.2
safetensors==0.4.2
salesforce-lavis==1.0.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.13.1 
selfies==2.1.1
sentencepiece==0.2.0
sentry-sdk==2.14.0
setproctitle==1.3.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.38.5
starsessions==1.3.0
streamlit==1.32.0
sympy==1.12
tenacity==8.2.3
tensorboardX==2.6.2.2
text-unidecode==1.3
thinc==8.2.3
threadpoolctl==3.3.0
tifffile==2023.7.10
timm==0.4.12
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1+cu118
torch_geometric==2.5.0
torchaudio==2.3.1+cu118
torchdata==0.7.1
torchmetrics==1.3.1
torchvision==0.18.1+cu118
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.44.2
triton==2.3.1
typer==0.9.0
types-python-dateutil==2.9.0.20240906
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.20
uvicorn==0.30.6
virtualenv==20.25.1
wandb==0.17.9
wasabi==1.1.2
watchdog==4.0.0
wcwidth==0.2.13
weasel==0.3.4
web.py==0.62
webdataset==0.2.86
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
xarray==2023.1.0
xmod==1.8.1
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
zipp==3.17.0
